requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
gunicorn
flask-cors
PyJWT
backtrader
pandas
matplotlib
